added new story for accordion header icon
