+++ src/components/Accordion/AccordionHeaderIcon.stories.tsx
@@ -1,0 +1,26 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import React from "react";
+import { Meta, Story } from "@storybook/react";
+import { AccordionHeaderIcon as AccordionHeaderIconComponent } from "./AccordionHeaderIcon";
+import { AccordionHeaderIconProps, AccordionHeaderIconSize } from "./types";
+
+// eslint-disable-next-line import/no-default-export
+export default {
+    title: "Components/Accordion/Accordion Header Icon",
+    component: AccordionHeaderIconComponent,
+    argTypes: {
+        size: { control: "select", options: Object.values(AccordionHeaderIconSize) },
+    },
+    args: {
+        isOpen: false,
+        disabled: false,
+        size: AccordionHeaderIconSize.Medium,
+    },
+} as Meta<AccordionHeaderIconProps>;
+
+export const AccordionHeaderIcon: Story<AccordionHeaderIconProps> = (args) => (
+    <div className="tw-flex tw-items-center">
+        <AccordionHeaderIconComponent {...args} />
+    </div>
+);
