+++ src/components/Accordion/index.ts
@@ -1,3 +1,4 @@
 export * from "./Accordion";
 export * from "./AccordionHeader";
+export * from "./AccordionHeaderIcon";
 export * from "./types";
